+++ src/app/admin/manage-restaurant/page.tsx
@@ -2,8 +2,10 @@ import React, { useEffect, useState } from "react";
 // import AdminButton from "../manage-restaurant/adminButton";
 import RestaurantList from "../../home/restaurantItem";
 import Link from "next/link";
+import router from "next/router";
 
 interface Restaurant {
+  id : number;
   name: string;
   content: string;
   thumbnail : string
@@ -11,6 +13,13 @@ interface Restaurant {
 
 const Restaurant: React.FC = () => {
   const user = "admin"; 
+
+  useEffect(()=>{
+    if(user !=="admin"){
+      alert("관리자가 아닙니다.")
+      router.push("/")
+    }
+  })
 
   return (
     <div className="h-screen w-screen flex justify-evenly flex-wrap gap-5 mt-[4%]">
